requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt
channels
Pillow
channels_redis
factory_boy
daphne
# brew
redis 